fix: use injected clocks for query timing
