+++ pkg/services/worker_sessions/internal/service/continue.go
@@ -9,7 +9,6 @@ import (
 	"slices"
 	"sort"
 	"strings"
-	"time"
 
 	"github.com/portpowered/infinite-you/pkg/services/events"
 	providersessions "github.com/portpowered/infinite-you/pkg/services/provider_sessions"
@@ -687,16 +686,16 @@ func (r *registry) ListWorkerSessionObservations(
 	ctx context.Context,
 	req workersessions.ListWorkerSessionObservationsRequest,
 ) (workersessions.ListWorkerSessionObservationsResult, error) {
-	listStartedAt := time.Now()
+	listStartedAt := r.clock.Now()
 	query, err := r.parseObservationListQuery(ctx, req)
 	if err != nil {
 		return workersessions.ListWorkerSessionObservationsResult{}, err
 	}
-	idCollectionStartedAt := time.Now()
+	idCollectionStartedAt := r.clock.Now()
 	ids := r.observationListIDs(query.cursor, query.scope, req.States)
-	idCollectionDuration := time.Since(idCollectionStartedAt)
+	idCollectionDuration := r.clock.Now().Sub(idCollectionStartedAt)
 	pageIDs := observationListPage(ids, query.limit)
-	projectionStartedAt := time.Now()
+	projectionStartedAt := r.clock.Now()
 	observations, err := r.projectObservationList(ctx, pageIDs)
 	if err != nil {
 		return workersessions.ListWorkerSessionObservationsResult{}, err
@@ -708,8 +707,8 @@ func (r *registry) ListWorkerSessionObservations(
 		"candidate_count", len(ids),
 		"result_count", len(observations),
 		"id_collection_duration_ms", idCollectionDuration.Milliseconds(),
-		"projection_duration_ms", time.Since(projectionStartedAt).Milliseconds(),
-		"total_duration_ms", time.Since(listStartedAt).Milliseconds(),
+		"projection_duration_ms", r.clock.Now().Sub(projectionStartedAt).Milliseconds(),
+		"total_duration_ms", r.clock.Now().Sub(listStartedAt).Milliseconds(),
 	)
 	r.logger.Info("worker session top-level observation list", "scope", string(query.scope), "state_count", len(req.States), "result_count", len(observations), "has_next", nextToken != "")
 	return workersessions.ListWorkerSessionObservationsResult{
