+++ pkg/services/worker_sessions/internal/service/invoke_session.go
@@ -500,7 +500,7 @@ func (r *registry) finishObservationLocked(id string, endedAt time.Time) {
 }
 
 func (r *registry) ListObservations(ctx context.Context, req workersessions.ListObservationsRequest) (workersessions.ListObservationsResult, error) {
-	listStartedAt := time.Now()
+	listStartedAt := r.clock.Now()
 	if err := req.Validate(); err != nil {
 		r.logger.Info("worker session observation list rejected", "outcome", "invalid")
 		return workersessions.ListObservationsResult{}, err
@@ -521,14 +521,14 @@ func (r *registry) ListObservations(ctx context.Context, req workersessions.List
 		}
 	}
 	r.mu.RUnlock()
-	idCollectionDuration := time.Since(listStartedAt)
+	idCollectionDuration := r.clock.Now().Sub(listStartedAt)
 	if len(ids) == 0 {
 		r.logger.Info("worker session observation list", "workID", req.WorkID, "outcome", "not_found")
 		return workersessions.ListObservationsResult{}, workersessions.ErrObservationWorkNotFound
 	}
 	sortObservationOrder(ids)
 
-	projectionStartedAt := time.Now()
+	projectionStartedAt := r.clock.Now()
 	observations := make([]workersessions.Observation, 0, len(ids))
 	for _, item := range ids {
 		projected, err := r.projectObservation(ctx, item.id)
@@ -543,8 +543,8 @@ func (r *registry) ListObservations(ctx context.Context, req workersessions.List
 		"candidate_count", len(ids),
 		"result_count", len(observations),
 		"id_collection_duration_ms", idCollectionDuration.Milliseconds(),
-		"projection_duration_ms", time.Since(projectionStartedAt).Milliseconds(),
-		"total_duration_ms", time.Since(listStartedAt).Milliseconds(),
+		"projection_duration_ms", r.clock.Now().Sub(projectionStartedAt).Milliseconds(),
+		"total_duration_ms", r.clock.Now().Sub(listStartedAt).Milliseconds(),
 	)
 	r.logger.Info("worker session observation list", "workID", req.WorkID, "outcome", "success", "result_count", len(observations))
 	return workersessions.ListObservationsResult{Observations: observations}, nil
